requirements.txt
pytest
selene>=2.0.0b14
allure-pytest